requirements.txt
spotipy
openai
python-dotenv
transformers
